requirements.txt
asgiref==3.4.1
autopep8==1.7.0
beautifulsoup4==4.11.1
Django==3.2.16
django-bootstrap-v5==1.0.11
django-geojson==3.2.1
django-leaflet==0.28.3
et-xmlfile==1.1.0
fontawesomefree==6.2.0
importlib-metadata==4.8.3
openpyxl==3.0.10
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pytz==2022.4
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.1.1
zipp==3.6.0
